+++ packages/freezed/lib/src/freezed_generator.dart
@@ -326,14 +326,11 @@ class FreezedGenerator extends ParserGenerator<_GlobalData, Data, Freezed> {
 
     final needsJsonSerializable = globalData.hasJson &&
         element.constructors.any((element) {
-          if (element.isFactory && element.name == 'fromJson') {
-            final ast = element.session
-                .getParsedLibraryByElement(element.library)
-                .getElementDeclaration(element)
-                ?.node;
-            return ast.endToken.stringValue == ';';
-          }
-          return false;
+          return element.isFactory &&
+              element.name == 'fromJson' &&
+              element.parameters.length == 1 &&
+              (element.parameters[0].type.getDisplayString(withNullability: false) == 'dynamic' ||
+                  element.parameters[0].type.getDisplayString(withNullability: false) == 'Map<String, dynamic>');
         });
 
     return Data(